+++ src/main/java/com/codelog/fitch/math/Vector4.java
@@ -1,0 +1,32 @@
+package com.codelog.fitch.math;
+
+@SuppressWarnings("WeakerAccess")
+public class Vector4 {
+
+    public double x;
+    public double y;
+    public double z;
+    public double w;
+
+    public Vector4(double x, double y, double z, double w) {
+        this.x = x;
+        this.y = y;
+        this.z = z;
+        this.w = w;
+    }
+
+    public Vector4(float x, float y, float z, float w) {
+        this.x = x;
+        this.y = y;
+        this.z = z;
+        this.w = w;
+    }
+
+    public Vector4(int x, int y, int z, int w) {
+        this.x = x;
+        this.y = y;
+        this.z = z;
+        this.w = w;
+    }
+
+}